api/requirements.txt
amqp==2.5.2
apipkg==1.5
appdirs==1.4.3
attrs==19.3.0
Babel==2.8.0
billiard==3.6.2.0
cairocffi
CairoSVG
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.8
cssselect2==0.2.2
debtcollector==2.0.0
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.13
django-appconf==1.0.3
django-celery-beat==1.6.0
django-celery-email==3.0.0
django-celery-results==1.2.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.8
django-filter==2.2.0
django-imagekit==4.0.2
django-rest-swagger==2.2.0
django-storage-swift==1.2.19
django-storages==1.9.1
django-timezone-field==3.1
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
djangorestframework-gis==0.15
djangorestframework-xml==1.4.0
drf-amsterdam==0.3.2
drf-extensions==0.6.0
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
entrypoints==0.3
execnet==1.7.1
factory-boy==2.12.0
Faker==4.0.0
filelock==3.0.12
flake8==3.7.9
flex==6.14.1
freezegun==0.3.14
graphene==2.1.8
graphene-django==2.10.1
graypy==2.1.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.5.0
iso8601==0.1.12
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
jsonpointer==2.0
jsonschema==3.2.0
jwcrypto==0.6.0
keystoneauth1==3.18.0
kombu==4.6.7
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
msgpack==0.6.2
netaddr==0.7.19
netifaces==0.10.9
openapi-codec==1.3.2
os-service-types==1.7.0
oslo.config==8.0.0
oslo.i18n==4.0.0
oslo.serialization==3.0.0
oslo.utils==4.0.0
packaging==20.1
pbr==5.4.4
pilkit==2.0
Pillow==7.1.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pygelf==0.3.6
PyJWT==1.7.1
pyparsing==2.4.6
Pyphen==0.9.5
pyrsistent==0.15.7
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-forked==1.1.3
pytest-timeout==1.3.4
pytest-xdist==1.31.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-keystoneclient==3.22.0
python-magic==0.4.15
python-swiftclient==3.9.0
pytz==2019.3
PyYAML==5.3
raven==6.10.0
requests==2.22.0
requests-mock==1.5.2
rfc3986==1.3.2
rfc3987==1.3.8
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
stevedore==1.32.0
strict-rfc3339==0.7
text-unidecode==1.3
tinycss2==1.0.2
toml==0.10.0
tox==3.14.5
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
validate-email==1.3
vine==1.3.0
wcwidth==0.1.8
WeasyPrint==51
webencodings==0.5.1
wrapt==1.12.0
xmlunittest==0.5.0
zipp==2.2.0
